asg6

+++ asg6/manipulation.h
@@ -16,10 +16,10 @@ class Manipulation
 {
 public:
     static ImageData* verticalFlip(ImageData* inputImage);
-    static void translate(Matrix3x3 &M, int dx, int dy);
+    static void translate(Matrix3x3 &M, float dx, float dy);
     static void rotate(Matrix3x3 &M, float theta);
-    static void shear(Matrix3x3 &M, int shx, int shy); 
-    static void scale(Matrix3x3 &M, int sx, int sy);
+    static void shear(Matrix3x3 &M, float shx, float shy); 
+    static void scale(Matrix3x3 &M, float sx, float sy);
     static ImageData* warper(ImageData* inputImage, Matrix3x3 &M);
     static ImageData* twirl(ImageData* inputImage, float s, float cx, float cy);
